pyproject.toml
[project]
name = "api-base-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "cryptography>=48.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.136.1",
    "psycopg2-binary>=2.9.12",
    "pydantic>=2.13.4",
    "pydantic-settings>=2.14.0",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.27",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint.per-file-ignores]
"alembic/versions/*.py" = ["F401"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
strict = true


[dependency-groups]
dev = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "mypy>=2.0.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.12",
    "watchfiles>=1.1.1",
]


